refactor: Line과 관련된 Sections, Section, Distance 엔티티 개선

+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -27,7 +27,7 @@ public class Line extends BaseEntity {
         this.color = color;
     }
 
-    public Line(String name, String color, Station upStation, Station downStation, int distance) {
+    public Line(String name, String color, Station upStation, Station downStation, long distance) {
         this.name = name;
         this.color = color;
         sections = new Sections(createSection(upStation, downStation, distance));
@@ -65,7 +65,7 @@ public class Line extends BaseEntity {
         return sections.getStations();
     }
 
-    private Section createSection(Station upStation, Station downStation, int distance) {
+    private Section createSection(Station upStation, Station downStation, long distance) {
         return new Section(this, upStation, downStation, distance);
     }
 }
